Removing ui footer component, changing main footer to footer.
