+++ core/app/assets/styles/app-imports.scss
@@ -24,9 +24,8 @@
 @import "../../modules/import-export/assts/js/pages/import-success/import-success-list/import-success-list.scss";
 @import "../../modules/import-export/assts/js/ui/heading/heading.scss";
 @import "../../modules/import-export/assts/js/ui/message/message.scss";
-@import "../../modules/import-export/assts/js/shared/main-footer/main-footer.scss";
 @import "../../modules/import-export/assts/js/ui/text/text.scss";
 @import "../../modules/import-export/assts/js/shared/kit-content-list/kit-content-list.scss";
 @import "../../modules/import-export/assts/js/ui/list/list.scss";
-@import "../../modules/import-export/assts/js/ui/footer/footer.scss";
 @import "../../modules/import-export/assts/js/shared/kit-content-list/kit-content-select/kit-content-select.scss";
+@import "../../modules/import-export/assts/js/shared/footer/footer.scss";
