+++ core/app/modules/import-export/assts/js/shared/footer/footer.scss
@@ -4,6 +4,11 @@
 	left: 0;
 	bottom: 0;
 	width: 100%;
-	border-top: 1px solid tints(300);
 	padding: spacing(8);
+	background-color: tints(200);
+
+	&__separator {
+		border-top: 1px solid tints(300);
+	}
 }
+//
