--- core/app/modules/import-export/assts/js/shared/main-footer/main-footer.scss
@@ -1,4 +1,0 @@
-.import-export-footer {
-	padding: spacing(8);
-	background-color: tints(200);
-}
