--- core/app/modules/import-export/assts/js/ui/footer/footer.js
@@ -1,13 +1,0 @@
-import './footer.scss';
-
-export default function Footer( props ) {
-	return (
-		<footer className="import-export-footer">
-			{ props.children }
-		</footer>
-	);
-}
-
-Footer.propTypes = {
-	children: PropTypes.object,
-};
